# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BPUT Feedback Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /faculty:
    get:
      operationId: listFaculty
      tags: [faculty]
      summary: List faculty members
      parameters:
        - name: departmentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of faculty
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Faculty"
    post:
      operationId: createFaculty
      tags: [faculty]
      summary: Create faculty member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFacultyBody"
      responses:
        "201":
          description: Created faculty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faculty"
  /faculty/{id}:
    get:
      operationId: getFaculty
      tags: [faculty]
      summary: Get faculty by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Faculty details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faculty"

  # ── Courses ────────────────────────────────────────────────────────────────
...
  /feedback:
    get:
      operationId: listFeedback
      tags: [feedback]
      summary: List feedback entries
      parameters:
        - name: courseId
          in: query
          schema:
            type: integer
        - name: departmentId
          in: query
          schema:
            type: integer
        - name: facultyId
          in: query
          schema:
            type: integer
        - name: semester
          in: query
          schema:
            type: integer
        - name: academicYear
          in: query
          schema:
            type: string
        - name: feedbackType
          in: query
          schema:
            type: string
            enum: [semester_end, mid_semester, event_based, placement]
      responses:
        "200":
          description: List of feedback
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeedbackEntry"
    post:
      operationId: submitFeedback
      tags: [feedback]
      summary: Submit feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitFeedbackBody"
      responses:
        "201":
          description: Submitted feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackEntry"
  /feedback/{id}:
    get:
      operationId: getFeedback
      tags: [feedback]
      summary: Get feedback by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Feedback details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackEntry"

  # ── Feedback Windows ───────────────────────────────────────────────────────
  /windows:
    get:
      operationId: listWindows
      tags: [windows]
      summary: List feedback collection windows
      responses:
        "200":
          description: List of windows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeedbackWindow"
    post:
      operationId: createWindow
      tags: [windows]
      summary: Create feedback window
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWindowBody"
      responses:
        "201":
          description: Created window
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackWindow"
  /windows/{id}:
    get:
      operationId: getWindow
      tags: [windows]
      summary: Get window by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Window details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackWindow"
    patch:
      operationId: updateWindow
      tags: [windows]
      summary: Update feedback window
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWindowBody"
      responses:
        "200":
          description: Updated window
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackWindow"

  # ── Analytics ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Faculty:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        designation:
          type: string
        departmentId:
          type: integer
        departmentName:
          type: string
        avgRating:
          type: number
          format: float
        totalFeedbackCount:
          type: integer
      required: [id, name, designation, departmentId]
    CreateFacultyBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        designation:
          type: string
        departmentId:
          type: integer
      required: [name, designation, departmentId]
...
    FeedbackEntry:
      type: object
      properties:
        id:
          type: integer
        referenceId:
          type: string
        courseId:
          type: integer
        courseName:
          type: string
        courseCode:
          type: string
        facultyId:
          type: integer
        facultyName:
          type: string
        departmentId:
          type: integer
        departmentName:
          type: string
        semester:
          type: integer
        academicYear:
          type: string
        studentYear:
          type: integer
        section:
          type: string
        feedbackType:
          type: string
          enum: [semester_end, mid_semester, event_based, placement]
        ratingCourseContent:
          type: integer
          minimum: 1
          maximum: 5
        ratingTeachingQuality:
          type: integer
          minimum: 1
          maximum: 5
        ratingLabFacilities:
          type: integer
          minimum: 1
          maximum: 5
        ratingStudyMaterial:
          type: integer
          minimum: 1
          maximum: 5
        ratingOverall:
          type: integer
          minimum: 1
          maximum: 5
        comments:
          type: string
        isAnonymous:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - referenceId
        - courseId
        - departmentId
        - semester
        - academicYear
        - feedbackType
        - ratingCourseContent
        - ratingTeachingQuality
        - ratingLabFacilities
        - ratingStudyMaterial
        - ratingOverall
        - isAnonymous
        - createdAt
    SubmitFeedbackBody:
      type: object
      properties:
        courseId:
          type: integer
        facultyId:
          type: integer
        studentYear:
          type: integer
        section:
          type: string
        feedbackType:
          type: string
          enum: [semester_end, mid_semester, event_based, placement]
        ratingCourseContent:
          type: integer
          minimum: 1
          maximum: 5
        ratingTeachingQuality:
          type: integer
          minimum: 1
          maximum: 5
        ratingLabFacilities:
          type: integer
          minimum: 1
          maximum: 5
        ratingStudyMaterial:
          type: integer
          minimum: 1
          maximum: 5
        ratingOverall:
          type: integer
          minimum: 1
          maximum: 5
        comments:
          type: string
        isAnonymous:
          type: boolean
      required:
        - courseId
        - feedbackType
        - ratingCourseContent
        - ratingTeachingQuality
        - ratingLabFacilities
        - ratingStudyMaterial
        - ratingOverall
    FeedbackWindow:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        feedbackType:
          type: string
          enum: [semester_end, mid_semester, event_based, placement]
        academicYear:
          type: string
        semester:
          type: integer
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        isActive:
          type: boolean
        departmentIds:
          type: array
          items:
            type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, title, feedbackType, academicYear, semester, startDate, endDate, isActive, createdAt]
    CreateWindowBody:
      type: object
      properties:
        title:
          type: string
        feedbackType:
          type: string
          enum: [semester_end, mid_semester, event_based, placement]
        academicYear:
          type: string
        semester:
          type: integer
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        departmentIds:
          type: array
          items:
            type: integer
      required: [title, feedbackType, academicYear, semester, startDate, endDate]
    UpdateWindowBody:
      type: object
      properties:
        title:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        isActive:
          type: boolean
...