# lib/api-zod/src/generated/types/createWindowBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BPUT Feedback Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateWindowBodyFeedbackType } from "./createWindowBodyFeedbackType";

export interface CreateWindowBody {
  title: string;
  feedbackType: CreateWindowBodyFeedbackType;
  academicYear: string;
  semester: number;
  startDate: Date;
  endDate: Date;
  departmentIds?: number[];
}