# lib/api-zod/src/generated/types/feedbackWindow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BPUT Feedback Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackWindowFeedbackType } from "./feedbackWindowFeedbackType";

export interface FeedbackWindow {
  id: number;
  title: string;
  feedbackType: FeedbackWindowFeedbackType;
  academicYear: string;
  semester: number;
  startDate: Date;
  endDate: Date;
  isActive: boolean;
  departmentIds?: number[];
  createdAt: Date;
}