# lib/api-zod/src/generated/types/submitFeedbackBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BPUT Feedback Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { SubmitFeedbackBodyFeedbackType } from "./submitFeedbackBodyFeedbackType";

export interface SubmitFeedbackBody {
  courseId: number;
  facultyId?: number;
  studentYear?: number;
  section?: string;
  feedbackType: SubmitFeedbackBodyFeedbackType;
  /**
   * @minimum 1
   * @maximum 5
   */
  ratingCourseContent: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  ratingTeachingQuality: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  ratingLabFacilities: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  ratingStudyMaterial: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  ratingOverall: number;
  comments?: string;
  isAnonymous?: boolean;
}